refactor: Package 구조 변경 - info 추가
