+++ src/test/java/jwp/was/webapplicationserver/configure/controller/info/ModelAndViewTest.java
@@ -1,4 +1,4 @@
-package jwp.was.webapplicationserver.configure.controller;
+package jwp.was.webapplicationserver.configure.controller.info;
 
 import static com.google.common.net.HttpHeaders.CONTENT_TYPE;
 import static org.assertj.core.api.Assertions.assertThat;
