+++ src/test/java/jwp/was/webapplicationserver/controller/UserPageControllerTest.java
@@ -3,7 +3,7 @@ package jwp.was.webapplicationserver.controller;
 import static org.assertj.core.api.Assertions.assertThat;
 
 import jwp.was.webapplicationserver.configure.ConfigureMaker;
-import jwp.was.webapplicationserver.configure.controller.ModelAndView;
+import jwp.was.webapplicationserver.configure.controller.info.ModelAndView;
 import jwp.was.webserver.HttpMethod;
 import jwp.was.webserver.HttpStatusCode;
 import jwp.was.webserver.dto.HttpRequest;
